Remove Eucalyptus support
