+++ src/main/resources/hudson/plugins/ec2/EC2Cloud/help-region.html
