--- src/main/resources/hudson/plugins/ec2/EC2Cloud/help-s3EndpointUrl.html
@@ -1,33 +1,0 @@
-<!--
-The MIT License
-
-Copyright (c) 2004-, Kohsuke Kawaguchi, Sun Microsystems, Inc., and a number of other of contributors
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
--->
-<div>
-    This field is optional. It controls what S3 API endpoint to connect to. 
-    By default https://s3.amazonaws.com/ is used (note that this works with all
-    amazonaws zones - s3 transparently redirects between zones).
-
-    When using Ubuntu Enterprise Cloud (Eucalyptus) the default URL should be
-    something like http://cluster-controller-ip:8773/services/Walrus.
-
-    For other EC2/S3 compatible servers please see your product documentation.
-</div>
