+++ src/main/resources/hudson/plugins/ec2/Messages.properties
@@ -8,6 +8,6 @@ EC2SpotSlave.AmazonEC2SpotInstance=Amazon EC2 Spot Instance
 EC2SpotSlave.Spot1=Spot $
 EC2SpotSlave.Spot2= max bid price
 
-AmazonEC2Cloud.NonUniqName=Cloud name must be unique across EC2 clouds
-AmazonEC2Cloud.MalformedUrl=The URL is malformed. The default endpoint will be used
+EC2Cloud.NonUniqName=Cloud name must be unique across EC2 clouds
+EC2Cloud.MalformedUrl=The URL is malformed. The default endpoint will be used
 General.MissingPermission=You do not have the Overall/Administer right to modify this field
