+++ src/test/java/hudson/plugins/ec2/ConfigurationAsCodeTest.java
@@ -32,7 +32,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("EC2CloudEmpty.yml")
     public void testEmptyConfig() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("empty");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("empty");
         assertNotNull(ec2Cloud);
         assertEquals(0, ec2Cloud.getTemplates().size());
     }
@@ -40,7 +40,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("UnixData.yml")
     public void testUnixData() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("production");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("production");
         assertNotNull(ec2Cloud);
         assertTrue(ec2Cloud.isUseInstanceProfileForCredentials());
 
@@ -77,7 +77,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("Unix.yml")
     public void testUnix() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("staging");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("staging");
         assertNotNull(ec2Cloud);
         assertTrue(ec2Cloud.isUseInstanceProfileForCredentials());
 
@@ -99,7 +99,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("WindowsData.yml")
     public void testWindowsData() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("development");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("development");
         assertNotNull(ec2Cloud);
         assertTrue(ec2Cloud.isUseInstanceProfileForCredentials());
 
@@ -128,7 +128,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("BackwardsCompatibleConnectionStrategy.yml")
     public void testBackwardsCompatibleConnectionStrategy() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("us-east-1");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("us-east-1");
         assertNotNull(ec2Cloud);
 
         final List<SlaveTemplate> templates = ec2Cloud.getTemplates();
@@ -162,7 +162,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("Unix-withMinimumInstancesTimeRange.yml")
     public void testConfigAsCodeWithMinimumInstancesTimeRange() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("timed");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("timed");
         assertNotNull(ec2Cloud);
         assertTrue(ec2Cloud.isUseInstanceProfileForCredentials());
 
@@ -191,7 +191,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("Ami.yml")
     public void testAmi() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("test");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("test");
         assertNotNull(ec2Cloud);
 
         final List<SlaveTemplate> templates = ec2Cloud.getTemplates();
@@ -224,7 +224,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("MacData.yml")
     public void testMacData() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("production");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("production");
         assertNotNull(ec2Cloud);
         assertTrue(ec2Cloud.isUseInstanceProfileForCredentials());
 
@@ -254,7 +254,7 @@ public class ConfigurationAsCodeTest {
     @Test
     @ConfiguredWithCode("Mac.yml")
     public void testMac() throws Exception {
-        final AmazonEC2Cloud ec2Cloud = (AmazonEC2Cloud) Jenkins.get().getCloud("staging");
+        final EC2Cloud ec2Cloud = (EC2Cloud) Jenkins.get().getCloud("staging");
         assertNotNull(ec2Cloud);
         assertTrue(ec2Cloud.isUseInstanceProfileForCredentials());
