+++ src/test/java/hudson/plugins/ec2/EC2SlaveMonitorTest.java
@@ -72,7 +72,7 @@ public class EC2SlaveMonitorTest {
                 EC2AbstractSlave.DEFAULT_METADATA_HOPS_LIMIT,
                 EC2AbstractSlave.DEFAULT_METADATA_SUPPORTED);
         SSHCredentialHelper.assureSshCredentialAvailableThroughCredentialProviders("ghi");
-        AmazonEC2Cloud cloud = new AmazonEC2Cloud(
+        EC2Cloud cloud = new EC2Cloud(
                 "us-east-1",
                 true,
                 "abc",
@@ -142,7 +142,7 @@ public class EC2SlaveMonitorTest {
                 EC2AbstractSlave.DEFAULT_METADATA_HOPS_LIMIT,
                 EC2AbstractSlave.DEFAULT_METADATA_SUPPORTED);
         SSHCredentialHelper.assureSshCredentialAvailableThroughCredentialProviders("ghi");
-        AmazonEC2Cloud cloud = new AmazonEC2Cloud(
+        EC2Cloud cloud = new EC2Cloud(
                 "us-east-1",
                 true,
                 "abc",
