+++ src/test/java/hudson/plugins/ec2/TemplateLabelsTest.java
@@ -42,7 +42,7 @@ public class TemplateLabelsTest {
     @Rule
     public JenkinsRule r = new JenkinsRule();
 
-    private AmazonEC2Cloud ac;
+    private EC2Cloud ac;
     private final String LABEL1 = "label1";
     private final String LABEL2 = "label2";
 
@@ -105,7 +105,7 @@ public class TemplateLabelsTest {
         List<SlaveTemplate> templates = new ArrayList<>();
         templates.add(template);
 
-        ac = new AmazonEC2Cloud("us-east-1", false, "abc", "us-east-1", "ghi", null, "3", templates, null, null);
+        ac = new EC2Cloud("us-east-1", false, "abc", "us-east-1", "ghi", null, "3", templates, null, null);
     }
 
     @Test
